requirements.txt
streamlit==1.49.1
pandas==2.3.2
reportlab==4.4.3
python-dotenv==1.1.1